perf(clausura): picks rivales en int16 — el transitorio mas grande del constructor
`sample_picks_match` devolvia (R, S) int64 y el camino i.i.d. hacia lo mismo con
rng.choice. Son indices de marcador (0..35): a 19.200 sorteos y 715 rivales eran
110 MB por partido donde alcanzan 27. numpy indexa igual con int16.

Verificado identico en los dos caminos (i.i.d. y RivalModel): mismo E[premio] y mismo
rivals_total.sum() con la misma semilla.

De paso corrige el docstring de _tilted_sample, que despues del troceo por rivales
seguia diciendo que se troceaba por simulaciones y que u se sorteaba entero — o sea
describiendo la version anterior.

+++ src/clausura/economics.py
@@ -211,7 +211,8 @@ class SeasonSimulator:
             self.rivals_total = self._rivals_total + rivals.residuo.astype(np.int32)[:, None]
         else:
             for m in range(self.n_matches):
-                rp = rng.choice(N_SCORES, size=(R, S), p=pool_q[m])
+                # int16: índices de marcador, 110 MB → 27 por partido a S=19.200
+                rp = rng.choice(N_SCORES, size=(R, S), p=pool_q[m]).astype(np.int16)
                 pts = self.pm[m][rp, self.actual[m][None, :]]
                 _acumular(m, pts)
             self.rivals_total = self._rivals_total
